README.md
ATLASSIAN

I choose Atlassian JIRA and Confluence for my presentation as it is commonly used and popular in market. It helps me learn the tools and an advantage for career. It helps most of the teams work together and manage projects in workspaces. Its a combination of issue tracking and agile processes of JIRA with the knowledge sharing and documentation features of Confluence.

JIRA

https://www.atlassian.com/software/jira

It gives teams a single place to plan, track, and organize their work in an efficient way. With flexible processes, agile boards, and a lot of ways to report, JIRA makes it easy for people to work together and speeds up the software development lifecycle. It has many useful tools, such as bug tracking, version control integration, and real-time analytics, which help teams work faster and make better products. JIRA is good for a wide range of businesses and team sizes because it can be changed and grown.


CONFLUENCE

https://www.atlassian.com/software/confluence

It is a powerful and flexible software tool for teamwork and documenting. It gives teams a central place to build, organize, and share information in an effective way. With tools like rich text editing, page hierarchy, and strong search, Confluence makes it easy for teams to collect, store, and share information in a way that is organized and easy to find. It makes teamwork easier by letting people respond in real time, keep track of versions. Confluence works well with other Atlassian tools like JIRA, making it easy for project management and documentation.