README.md
jenkins

https://ci.jenkins.io/

https://www.jenkins.io/

Jenkins is an open-source automation server that makes it easier to build, test, and deploy software projects. It also automates tasks that need to be done over and over again, so software can be released faster and more reliably. It enables continuous integration and delivery pipelines for teams of all sizes. Jenkins has a web-based interface that makes it easy configuration and monitoring of jobs. Its distributed design makes allows scalability and use resources efficiently. Jenkins has become a go-to tool for teams.


chocolatey

https://community.chocolatey.org/

Chocolatey is a user-friendly package manager for Windows that simplifies software installation and management. It has a large repository of pre-packaged software and a command-line interface that makes it easy to install and update software. Chocolatey makes it easier to handle software, which makes it a useful tool for Windows users for efficient software package management.