README.md
## A07 - WEB SCRAPING
### VINEELA SEERLA
### Description:

This project will combine a python GUI with a beautiful soup web scraper to: 1. grab, 2. save and 3. display the data.
The initial python gui will be used to enter the appropriate values to allow you to leverage the URL, meaning it will accept values for: month, day, year, airport code, and one of the following: daily, weekly, monthly.
The resulting python gui will display the received data in a tabular format.

### Files

|   #   | File            | Description                                          |
| :---: | --------------- | --------------------------------------------------   |
|   1   | gui.py          | gui layout to take input and scrap data gives gui o/p|
|   2   | input.png       | gui window that takes input values                   |
|   3   | output.png      | gui window that displays rendered values             |
|   4   | weather_data.csv| .csv file that stores rendered values                |

### Instructions

- Make sure you install library pysimplegui
bs4
selenium
selenium-wire
- Enter valid year-month-day and airport code and select filter
- Check the output data with actual URL data.

###  References

- https://data.library.virginia.edu/getting-started-with-web-scraping-in-python/
- https://pysimplegui.trinket.io/demo-programs#/tables/the-table-element
- https://www.pysimplegui.org/en/latest/